backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
pydantic>=2.0.0
Pillow>=10.0.0
opencv-python-headless>=4.9.0
numpy>=1.26.0
mediapipe>=0.10.0
deepface>=0.0.90
transformers>=4.40.0
torch>=2.0.0
torchvision>=0.15.0
google-generativeai>=0.5.0
python-dotenv>=1.0.0
httpx>=0.27.0
supabase>=2.3.0
